fix(parser): parse `i128` instead of `i64`

+++ compiler/lume_ast/src/lib.rs
@@ -1041,7 +1041,7 @@ impl Node for Literal {
 
 #[derive(Hash, Debug, Clone, PartialEq, Eq)]
 pub struct IntLiteral {
-    pub value: i64,
+    pub value: i128,
     pub location: Location,
     pub kind: Option<IntKind>,
 }
